Changes default to download in background and gives clearer description.
Changes the default to work the same as all modern programs would, without popping up endless modal boxes as the user uses the program. If the user wants to see the download progress with every pan and zoom action, the option should at least be worded so that a layman can understand what it means. I hope this change makes the checkbox easier to understand.